requirements.txt
urllib3==1.20
requests==2.10.0
gitpython
pyyaml==3.12
sqlite3
